refactor(animation): split 1,776-line file into 7 focused modules
Split monolithic animation/mod.rs into:
- tween.rs (307 lines): Tween, Animation, Animations presets
- keyframe.rs (360 lines): CssKeyframe, KeyframeAnimation, AnimationDirection, AnimationFillMode
- stagger.rs (102 lines): Stagger helper for animating multiple elements
- group.rs (144 lines): AnimationGroup, GroupMode for parallel/sequential animations
- choreographer.rs (103 lines): Choreographer for coordinating complex sequences
- presets.rs (203 lines): widget_animations module with common presets
- tests.rs (542 lines): All 70+ animation tests
- mod.rs (40 lines): Public exports

File reduced from 1,776 to 40 lines (modularized into 8 files).

Benefits:
- Single Responsibility: Each module has one clear purpose
- Easier navigation: Smaller, focused files
- Better testability: Tests separated into dedicated module
- Maintains backward compatibility: All public APIs unchanged
